requirements.txt
anytree==2.8.0
apyori==1.1.2
imbalanced_learn==0.8.1
matplotlib==3.4.1
numpy==1.20.0
orange3==3.31.1
pandas==1.3.4
pysmt==0.9.0
python-sat==0.1.7.dev14
scikit-learn==0.24.1
torch==1.10.1
torchvision==0.11.2
xgboost==1.5.1
z3-solver==4.8.17.0